move common functions
